Utils and SDK lib separation cleanup
