+++ lib/include/CL/SDK/Image.h
@@ -1,7 +1,10 @@
 #pragma once
 
 // OpenCL SDK includes
-#include "OpenCLUtils_Export.h"
+#include "OpenCLSDK_Export.h"
+
+// OpenCL includes
+#include <CL/cl.h>
 
 typedef struct cl_sdk_image
 {
@@ -10,8 +13,8 @@ typedef struct cl_sdk_image
 }
 cl_sdk_image;
 
-UTILS_EXPORT
+SDK_EXPORT
 cl_sdk_image cl_sdk_read_image(const char* file_name, cl_int* err);
 
-UTILS_EXPORT
+SDK_EXPORT
 void cl_sdk_write_image(const char * file_name, const cl_sdk_image * im, cl_int * err);
