+++ lib/include/CL/SDK/Image.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 // OpenCL SDK includes
-#include "OpenCLUtilsCpp_Export.h"
+#include "OpenCLSDKCpp_Export.h"
 
 // OpenCL includes
 #include <CL/opencl.hpp>
@@ -13,13 +13,13 @@ namespace sdk
     struct Image
     {
         int width, height, pixel_size;
-        std::vector<unsigned char> pixels;
+        cl::vector<unsigned char> pixels;
     };
 
-    UTILSCPP_EXPORT
+    SDKCPP_EXPORT
     Image read_image(const char* file_name, cl_int* err);
 
-    UTILSCPP_EXPORT
+    SDKCPP_EXPORT
     void write_image(const char* file_name, const Image& image, cl_int* err);
 }
 }
