+++ lib/include/CL/SDK/SDK.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+// OpenCL SDK includes
+#include "OpenCLSDK_Export.h"
+
+#include <CL/SDK/CLI.hpp>
+#include <CL/SDK/Image.hpp>
+#include <CL/SDK/InteropContext.hpp>
+#include <CL/SDK/InteropWindow.hpp>
+
+// OpenCL includes
+#include <CL/opencl.hpp>
