+++ lib/include/CL/Utils/Utils.hpp
@@ -1,9 +1,14 @@
 #pragma once
 
-// OpenCL SDK includes
+// OpenCL Utils includes
+#include "OpenCLUtils_Export.h"
+
 #include <CL/Utils/Detail.hpp>
 #include <CL/Utils/Error.hpp>
 #include <CL/Utils/Platform.hpp>
 #include <CL/Utils/Device.hpp>
 #include <CL/Utils/Context.hpp>
 #include <CL/Utils/Event.hpp>
+
+// OpenCL includes
+#include <CL/opencl.hpp>
