+++ lib/src/SDK/SDK.cpp
@@ -1,3 +1,6 @@
+// OpenCL SDK includes
+#include <CL/SDK/SDK.hpp>
+
 #include "Image.cpp"
 #include "InteropContext.cpp"
 #include "InteropWindow.cpp"
