+++ lib/src/Utils/Utils.cpp
@@ -1,4 +1,6 @@
+// OpenCL Utils includes
+#include <CL/Utils/Utils.hpp>
+
 #include "Platform.cpp"
 #include "Device.cpp"
 #include "Context.cpp"
-
